Resolves conflicts in main.py (both branches added routers), status_history/
step_version.py (both added FK, keep origin's ondelete=CASCADE), and
state_machine.py (combine origin's lazy-import circular-fix with this
branch's await-publish fix). Renumbered 006_add_pipeline_runs to
008_add_pipeline_runs since 006/007 were claimed by concurrent work
(006_add_scene_approvals, 007_add_step_versions_fk) while this branch was open.

Full suite after merge: 31 failed/248 passed (0 errors) -- healthier overall
baseline from other workers' concurrent fixes; this task's own 21 pipeline
tests all still green.

+++ backend/app/adapters/llm/embedding_gemini.py
@@ -152,4 +152,4 @@ def _parse_embed_response(raw: dict[str, Any]) -> Any:
             f"gemini_embedding: unexpected response shape: {exc}. Raw: {raw!r:.200}",
             retryable=True,
         ) from exc
-    return _np.array(values, dtype="float32")
+    return _np.array(values, dtype="float64")
